requirements.txt
Flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
deep-translator==1.11.4
Pillow==10.0.0
pyzbar==0.1.9
opencv-python-headless==4.8.0.76
numpy==1.24.3
gunicorn==21.2.0
